refactor: 重命名WatchedAniContext和WatchedAniContext为Provider，使名称和内容更匹配
